DEV: Export add emoji logic in textarea manipulation mixin
